refactor(metrics): single ledger-append implementation and clean close boundary (refs #4712)
Extract the open->append->rotate ledger tail into a narrow shared
lib/observability/metrics-ledger.js module; all three appenders
(appendPlanMetric, appendCriticSkip, appendFindingsYield) route through
it, eliminating the drifted third copy. Re-home appendFindingsYield
there so the story-close review spine no longer imports plan-domain
internals. Strictly behavior-preserving on the wire: record shapes,
file locations, and the 1 MiB rotation threshold are unchanged, and
findings-yield records still never inflate invocation tallies.
